Feature/stage

+++ src/main/java/searchengine/model/Lemma.java
@@ -17,6 +17,23 @@ import java.io.Serializable;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
+/**
+ * Сущность Lemma представляет лемму (основную форму слова),
+ * встречающуюся на сайте, и её частоту использования.
+ * <p>
+ * Класс привязан к таблице {@code lemma} в базе данных.
+ * Каждая лемма относится к конкретному сайту через связь Many-to-One.
+ * </p>
+ *
+ * <p>Пример использования:</p>
+ * <pre>
+ *     Lemma lemma = new Lemma();
+ *     lemma.setLemma("пример");
+ *     lemma.setFrequency(10);
+ * </pre>
+ *
+ * <p>Сравнение лемм происходит по частоте их появления.</p>
+ */
 @Entity
 @Table(name = "lemma")
 @Getter
@@ -26,18 +43,22 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 @ToString
 public class Lemma implements Serializable, Comparable<Lemma> {
 
+    /** Уникальный идентификатор леммы */
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private int id;
 
+    /** Сайт, на котором встречается лемма */
     @ManyToOne
     @JoinColumn(name = "site_id", nullable = false)
     private Site site;
 
-    @Column(name = "lemma",nullable = false)
+    /** Текстовое представление леммы */
+    @Column(name = "lemma", nullable = false)
     private String lemma;
 
+    /** Частота появления леммы на сайте */
     @Column(name = "frequency", nullable = false)
     private int frequency;
 
@@ -48,7 +69,6 @@ public class Lemma implements Serializable, Comparable<Lemma> {
         } else if (frequency < o.getFrequency()) {
             return -1;
         }
-
         return 0;
     }
 }
